% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{calculate_parameters}
\alias{calculate_parameters}
\title{Calculate parameters}
\usage{
calculate_parameters(hyper_parameters, restrictions)
}
\arguments{
\item{hyper_parameters}{An object of class \code{hyper_parameters}}

\item{restrictions}{An object of class \code{restrictions}}
}
\value{
An object of class \code{parameters}
}
\description{
Calculate parameters for simulating mendelian randomization (MR) data from the given hyper parameters and restrictions.
}
\details{
Additional constraints may be posed to the parameters values in order to garantee that the simulated data is valid.
}
