% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{calculate_summary_statistics}
\alias{calculate_summary_statistics}
\title{Calculate summary statistics}
\usage{
calculate_summary_statistics(sample)
}
\arguments{
\item{sample}{An object of class \code{sample}}
}
\value{
A data frame of \code{m + k} rows and \code{12} columns
}
\description{
Calculate summary statistics for a sample of mendelian randomization (MR) data.
}
\details{
The summary statistics are calculated for each SNP in the sample and include the following:
\itemize{
\item \code{beta_g_x}: The beta coefficient of the simple linear regression of between the SNP \code{g} and the exposure \code{x}
\item \code{beta_se_g_x}: The standard error of \code{beta_g_x}
\item \code{p_value_g_x}: The p-value of \code{beta_g_x}
\item \code{f_statistic_g_x}: The F-statistic of \code{beta_g_x}
\item \code{r2_g_x}: The R-squared of the simple linear regression of between the SNP \code{g} and the exposure \code{x}
\item \code{beta_g_y}: The beta coefficient of the simple linear regression of between the SNP \code{g} and the outcome \code{y}
\item \code{beta_se_g_y}: The standard error of \code{beta_g_y}
\item \code{p_value_g_y}: The p-value of \code{beta_g_y}
\item \code{f_statistic_g_y}: The F-statistic of \code{beta_g_y}
\item \code{r2_g_y}: The R-squared of the simple linear regression of between the SNP \code{g} and the outcome \code{y}
}
}
