% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{define_parameters}
\alias{define_parameters}
\title{Define parameters}
\usage{
define_parameters(
  m,
  k,
  p,
  alpha_u,
  sigma2_u,
  alpha_x,
  beta_g_x,
  beta_u_x,
  sigma2_x,
  alpha_y,
  beta_u_y,
  beta_x_y,
  sigma2_y
)
}
\arguments{
\item{m}{Number of non-zero effect SNPs. It should be a positive integer}

\item{k}{Number of zero effect SNPs. It should be a positive integer}

\item{p}{Minor allele frequency. It should be between \code{0} and \code{0.5}}

\item{alpha_u}{Intercept of U. It should be a number}

\item{sigma2_u}{Variance of U noise. It should be a positive number}

\item{alpha_x}{Intercept of X. It should be a number}

\item{beta_g_x}{Causal effect of G's on X. It should be a number}

\item{beta_u_x}{Causal effect of U on X. It should be a number}

\item{sigma2_x}{Variance of X noise. It should be a positive number}

\item{alpha_y}{Intercept of Y. It should be a number}

\item{beta_u_y}{Causal effect of U on Y. It should be a number}

\item{beta_x_y}{(Targeted causal effect) Causal effect of X on Y. It should be a number}

\item{sigma2_y}{Variance of Y noise. It should be a positive number}
}
\value{
An object of class \code{parameters}
}
\description{
Define parameters for simulating mendelian randomization (MR) data.
}
\details{
Additional constraints may be posed to the parameter values in order to garantee that the simulated data is valid.
}
