% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{define_restrictions}
\alias{define_restrictions}
\title{Define restrictions}
\usage{
define_restrictions()
}
\value{
An object of class \code{restrictions}
}
\description{
Define restrictions for simulating mendelian randomization (MR) data.
}
\details{
Additional constraints may be posed to the restrictions values in order to guarantee that the simulated data is valid.
}
