% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{generate_sample}
\alias{generate_sample}
\title{Generate sample}
\usage{
generate_sample(parameters, n)
}
\arguments{
\item{parameters}{An object of class \code{parameters}}

\item{n}{Sample size. It should be a positive integer}
}
\value{
An object of class \code{sample}
}
\description{
Generate a sample of size \code{n} from the model defined by \code{parameters}.
}
