% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beta_x_y.R
\name{get_beta_x_y}
\alias{get_beta_x_y}
\title{Get \code{beta_x_y}}
\usage{
get_beta_x_y(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Extra arguments. Currently not being used.}
}
\value{
A numeric value
}
\description{
Returns the causal effect of X on Y \code{beta_x_y}.
}
