% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endogenous_variables.R
\name{is_endogenous_variables}
\alias{is_endogenous_variables}
\title{Test if the object is \code{endogenous_variables}}
\usage{
is_endogenous_variables(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
Returns TRUE for \code{endogenous_variables} or subclasses thereof and FALSE for all other objects.
}
