% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous_variables.R
\name{is_exogenous_variables}
\alias{is_exogenous_variables}
\title{Test if the object is \code{exogenous_variables}}
\usage{
is_exogenous_variables(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
Returns TRUE for \code{exogenous_variables} or subclasses thereof and FALSE for all other objects.
}
