% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_parameters.R
\name{is_hyper_parameters}
\alias{is_hyper_parameters}
\title{Test if the object is \code{hyper_parameters}}
\usage{
is_hyper_parameters(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE if the object inherits from the \code{hyper_parameters} class, else FALSE
}
\description{
This function returns TRUE for \code{hyper_parameters} or subclasses thereof and FALSE for all other objects.
}
