% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{is_parameters}
\alias{is_parameters}
\title{Test if the object is \code{paremeters}}
\usage{
is_parameters(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
Returns TRUE for \code{paremeters} or subclasses thereof and FALSE for all other objects.
}
