% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{is_restrictions}
\alias{is_restrictions}
\title{Test if an object is of class \code{restrictions}}
\usage{
is_restrictions(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
Returns TRUE for \code{restrictions} or subclasses thereof and FALSE for all other objects.
}
