% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{is_sample}
\alias{is_sample}
\title{Test if the object is \code{sample}}
\usage{
is_sample(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
Returns TRUE for \code{sample} or subclasses thereof and FALSE for all other objects.
}
