% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_parameters.R
\name{print.hyper_parameters}
\alias{print.hyper_parameters}
\title{Print an object as \code{hyper_parameters}}
\usage{
\method{print}{hyper_parameters}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Extra arguments. Currently not being used.}
}
\description{
This function prints an object as if its class is \code{hyper_parameters}. It is intended to be used only for \code{hyper_parameters} or subclasses thereof.
}
